% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsglm.R
\name{fsglm}
\alias{fsglm}
\title{Execute single variable glm collectively.}
\usage{
fsglm(df, hier_key, y, x, family = "gaussian")
}
\arguments{
\item{df}{input data frame}

\item{hier_key}{group key column name with character type}

\item{y}{target column name with character type}

\item{x}{explanatory column name with character type}

\item{family}{glm family argument}
}
\description{
Execute single variable glm collectively.
}
